feat(stock): 添加板块详情信息接口和相关功能
- 在 StockService 中添加 sectorAll 方法
- 在 StockServiceImpl 中实现 sectorAll 方法，查询股票板块详情信息
- 在 StockController 中添加 sectorAll 接口
- 在 StockBlockRtInfoMapper 中添加 selectByDateTimeDesc 方法
- 更新相关 XML 文件，添加 SQL 查询语句
- 修改 SectorAllVO 中 tradeVol 类型为 BigDecimal

+++ stock-model/src/main/java/com/mozi/stock/vo/SectorAllVO.java
@@ -39,7 +39,7 @@ public class SectorAllVO {
   private String curDate;
 
   /** 成交总金额 */
-  private Long tradeVol;
+  private BigDecimal tradeVol;
 
   /** 涨跌幅（单位：百分比，如0.196表示0.196%） */
   private BigDecimal updownRate;
